+++ CHANGELOG.md
@@ -59,3 +59,5 @@ believed to be representative.
 - Added full set of CSS named colors
 - Fail more gracefully when `use` and `mask` elements have circular references
 - Minor cleanup:  Declare ScalableImage as `@immutable`
+- More forgiving use tag:  allow "href" instead of "xlink:href"
+- Fix bug with rrect when rx != ry